[test-stlport_xxx-exception] Don't run unwind1 for clang
This is done because clang after catching an exception
don't unwind stack to exactly the same position as it was before.
It's not completely clear is it bug or not, but it definitely not
the bug of NDK, since Linux/Darwin clang behave exactly the same.
In other words, even if it's bug, it's bug of Clang itself, and
irrelevant to NDK.

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>