[tests] Mark boost-dependencies test cases as broken for gcc-4.8
They are broken because Boost libraries were built with gcc-4.9 and
used GNU libstdc++-4.9 for compilation and linking. Thus, they contains
reference to std::__throw_out_of_range_fmt helper which was introduced
in GNU libstdc++-4.9 and absent in GNU libstdc++-4.8.
When we are building anything with gcc-4.8, it uses GNU libstdc++-4.8,
so these tests obviously will not link.

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>