[adbrunner] Include versioned SOs into list uploaded on device
This is fix of regression introduced in c6e6685fe; although, there
was another potential bug previously, which may include non-SO
files occasionally into list of binaries (liberty.sole, for example),
so this commit fixes regression as well as define wildcard for SO
libraries more explicitly - i.e. only 'lib*.so' and 'lib*.so.*' are
allowed now.

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>