[build/core] Fix bug #1188
There was wrong propagation of module's dependencies.
For example, having static library module (A) as dependency of shared
library module (S), it was propagated as direct dependency of all other
modules, which depends on S. As a result, static library was linked
to every module, resulting in multiple copies of static library and
its variables in the application.

To fix that, we only propagate "shared library" dependencies for
"shared library" modules, and don't touch "static library" dependencies
at all.

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>