cmake/modules/FindBoost.cmake: add a wrapper around real FindBoost.cmake with proper pathes
Our bundled Boost lies outside of the sysroot, so the default find logic
used in stock FindBoost.cmake does not find it. Help it a bit by pre-loading
certain internal variables with known valid pathes, then call the real routine
to do component detection and the rest of the boilerplate.